+++ app/src/main/java/ua/com/radiokot/photoprism/env/data/model/EnvConnectionParams.kt
@@ -1,22 +1,55 @@
 package ua.com.radiokot.photoprism.env.data.model
 
+import okhttp3.HttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrl
 
 /**
  * Holds parameters required to connect to the environment.
  */
 data class EnvConnectionParams(
-    val apiUrl: String,
+    /**
+     * Instance root URL.
+     *
+     * **May contain HTTP basic auth credentials**
+     *
+     * [What is Library root URL](https://github.com/Radiokot/photoprism-android-client/wiki/What-is-Library-root-URL%3F)
+     */
+    val rootUrl: HttpUrl,
+
+    /**
+     * Alias (name) of an installed client certificate to use for mTLS.
+     */
     val clientCertificateAlias: String?,
 ) {
-    companion object {
-        fun rootUrlToApiUrl(rootUrl: String): String =
-            rootUrl
-                .toHttpUrl()
-                .newBuilder()
-                .addPathSegment("api")
-                .addPathSegment("") // Force trailing slash.
-                .build()
-                .toString()
-    }
+    constructor(
+        rootUrlString: String,
+        clientCertificateAlias: String?
+    ) : this(
+        rootUrl = rootUrlString.toHttpUrl(),
+        clientCertificateAlias = clientCertificateAlias,
+    )
+
+    /**
+     * API root URL (not including the version).
+     *
+     * **May contain HTTP basic auth credentials**
+     */
+    val apiUrl: HttpUrl
+        get() = rootUrl
+            .newBuilder()
+            .addPathSegment("api")
+            .addPathSegment("") // Force trailing slash.
+            .build()
+
+    /**
+     * Library web client root URL.
+     *
+     * **May contain HTTP basic auth credentials**
+     */
+    val webLibraryUrl: HttpUrl
+        get() = rootUrl
+            .newBuilder()
+            .addPathSegment("library")
+            .addPathSegment("") // Force trailing slash.
+            .build()
 }